server_app/python/ai_service/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pillow
torch
torchvision
requests
python-multipart
